Lots of error and format fixing
uways - fixed multiple formatting errors and fixed some overlapping code.

since integrating the the front and back end the functionality has gone up by a lot

but we just have to figure out some errors that appear and test the website to make sure its running as it should.

we also have to sync the databases!